requirements.txt
flask
firebase-admin
python-docx
rapidfuzz
pymupdf==1.26.1
google-generativeai==0.4.0
openai>=1.3.7
python-dotenv>=1.0.1